Account for different type instances
